requirements.txt
pyalsaaudio
rpi-hardware-pwm
easing-functions
scipy
